// src/API.ts
export type CreateEventInput = {
  id?: string | null,
  name: string,
  description?: string | null,
  author: string,
  date: string,
};
// ...
export type UpdateEventInput = {
  id: string,
  name?: string | null,
  description?: string | null,
  author?: string | null,
  date?: string | null,
};